+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.2.35
 - Removed and reworked some legacy code
+- Fixed issue where secondary Damage formula wasn't working for Critial damage
 
 ## Version 0.2.34
 - Fixed an issue where rolling damage with no damage type would cause a diffrent issue some times.